Update style delete button.

+++ ozzang/client/src/components/views/MainPage/ClothesDetailPage.js
@@ -174,6 +174,10 @@ const ClothesDetailPage = () => {
   );
   const onDeleteButtonClick = useCallback(
     (event) => {
+      const clickDelete = window.confirm('정말로 삭제하시겠습니까?');
+      if (!clickDelete) {
+        return;
+      }
       deleteClothes(cloth);
       navigate("/main");
     },
